upgrade config

+++ zsh/zsh-config.sh
@@ -8,13 +8,27 @@ msg() {
 }
 
 oh_my_zsh() {
-    wget https://github.com/robbyrussell/oh-my-zsh/raw/master/tools/install.sh -O - | sh
+    if [[ ! -e $oh_my_zsh ]]; then
+        wget https://github.com/robbyrussell/oh-my-zsh/raw/master/tools/install.sh -O - | sh
+        plugins_install
+    else
+        printf "have oh-my-zsh directory on you $HOME."
+        printf "exit ..."
+        exit
+    fi
 }
 
 plugins_install() {
     if [ -d $oh_my_zsh/plugins ]; then
-        git clone git://github.com/zsh-users/zsh-autosuggestions $HOME/.oh-my-zsh/plugins/zsh-autosuggestions
-        git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+
+        if [ ! -e $oh_my_zsh/plugins/zsh-autosuggestions ]; then
+            git clone git://github.com/zsh-users/zsh-autosuggestions $HOME/.oh-my-zsh/plugins/zsh-autosuggestions
+        fi
+
+        if [ ! -e $oh_my_zsh/custom/plugins/zsh-syntax-highlighting ]; then
+            git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+        fi
+
     fi
 }
 
@@ -42,15 +56,6 @@ add_plugins_config() {
 ############################################################################################################## Main
 
 oh_my_zsh
-if [ -d $HOME/.oh-my-zsh ]; then
-    plugins_install
-else
-    msg "\n\n\t"
-    msg ".oh-my-zsh is not found"
-    msg "exit ..."
-    exit
-
-fi
 
 add_plugins_config
 config_link_files
